+++ include/boost/beast/http/file_body.hpp
@@ -34,4 +34,8 @@ using file_body = basic_file_body<file>;
 #include <boost/beast/http/impl/file_body_win32.hpp>
 #endif
 
+#ifndef BOOST_BEAST_NO_FILE_BODY_POSIX
+#include <boost/beast/http/impl/file_body_posix.hpp>
+#endif
+
 #endif